Fix tests to be runnable
One of the symlinks was still pointing to the old gundo file. Fixing
that wasn't enough. The tests were failing to even load due to some
problem with the symlinks (mundo's autoload functions were not defined
but mundo's commands were).

Remove all symlinks and add the paths in the test vimrc's runtimepath
instead.

Several tests fail (test-graph and test-preview). Not sure how long
they've been broken or if the breaks are valid.

(I found this because I was getting an error when grepping:

    ~/.vim/$ grep hsdfsf -R .
    grep: ./bundle/gundo/tests/vim_test/plugin/gundo.vim: No such file or directory

Deleting the symlink fixes that problem.)

I guess this means tests haven't been run in ages : )