Make help text fit in the default width
With the default width of 45 the help text doesn't quite fit. This
changes it so that it fits.

I tried to come up with a phrasing that was shorter so I didn't have the
remove the '-' characters, but I couldn't come up with anything. IMHO
it's also clear without them.